Refactor: taxa treemap into card and cell components
