+++ src/components/analytics/TaxaTreemap.tsx
@@ -1,7 +1,6 @@
 import { ResponsiveContainer, Tooltip, Treemap } from "recharts";
 
-import { Card, CardContent, CardDescription, CardHeader, CardTitle } from "@/components/ui/Card";
-import { Separator } from "@/components/ui/Separator";
+import { TaxaTreemapCell } from "@/components/analytics/TaxaTreemapCell";
 import { useFilters } from "@/lib/hooks/useFilters";
 import { useTaxaCount } from "@/lib/hooks/useTaxaCount";
 
@@ -22,45 +21,16 @@ interface TreemapItem {
   items?: OtherTaxaItem[];
 }
 
-const TREEMAP_COLORS = ["#8adf9f", "#62cd7b", "#3db85f", "#238f47", "#1b6c37"];
-const OTHER_TREEMAP_COLOR = "#5aa06f";
 const PRIMARY_TAXA_LIMIT = 8;
 const OTHER_PREVIEW_LIMIT = 8;
-const MIN_LABEL_WIDTH = 40;
-const MIN_LABEL_HEIGHT = 30;
-const MIN_COUNT_WIDTH = 72;
-const MIN_COUNT_HEIGHT = 44;
-const CELL_PADDING_X = 10;
-const PRIMARY_FONT_SIZE = 12;
-const SECONDARY_FONT_SIZE = 10;
 
 function getTreemapSize(count: number): number {
   return Math.log1p(count);
 }
 
-function truncateLabel(text: string, width: number, fontSize: number): string {
-  const availableWidth = width - CELL_PADDING_X * 2;
-  if (availableWidth <= 0) {
-    return "";
-  }
-
-  const averageCharWidth = fontSize * 0.58;
-  const maxChars = Math.floor(availableWidth / averageCharWidth);
-
-  if (maxChars <= 1) {
-    return "";
-  }
-
-  if (text.length <= maxChars) {
-    return text;
-  }
-
-  return `${text.slice(0, maxChars - 1)}\u2026`;
-}
-
 function TaxaTreemap({ deploymentId }: TaxaTreemapProps) {
   const { startDate, endDate, hub, taxonomyLevel, selectedTaxa, minConfidence } = useFilters();
-  const { data, isLoading } = useTaxaCount({
+  const { data, isError, isLoading } = useTaxaCount({
     start_time: startDate,
     end_time: endDate,
     device_id: hub ? [hub] : undefined,
@@ -71,21 +41,33 @@ function TaxaTreemap({ deploymentId }: TaxaTreemapProps) {
     sort_desc: true,
   });
 
-  let content: React.ReactNode;
+  const content = () => {
+    if (isLoading) {
+      return (
+        <div className="flex h-75 items-center justify-center">
+          <span className="text-sm text-muted-foreground">Loading chart...</span>
+        </div>
+      );
+    }
 
-  if (isLoading) {
-    content = (
-      <div className="flex h-75 items-center justify-center">
-        <span className="text-sm text-muted-foreground">Loading chart...</span>
-      </div>
-    );
-  } else if (!data || data.counts.length === 0) {
-    content = (
-      <div className="flex h-75 items-center justify-center">
-        <span className="text-sm text-muted-foreground">No taxonomy data for selected filters</span>
-      </div>
-    );
-  } else {
+    if (isError) {
+      return (
+        <div className="flex h-75 items-center justify-center">
+          <span className="text-sm text-destructive/80">Error loading taxonomy data.</span>
+        </div>
+      );
+    }
+
+    if (!data || data.counts.length === 0) {
+      return (
+        <div className="flex h-75 items-center justify-center">
+          <span className="text-sm text-muted-foreground">
+            No taxonomy data for selected filters
+          </span>
+        </div>
+      );
+    }
+
     const primaryTaxa = data.counts.slice(0, PRIMARY_TAXA_LIMIT);
     const otherTaxa = data.counts.slice(PRIMARY_TAXA_LIMIT);
 
@@ -105,7 +87,7 @@ function TaxaTreemap({ deploymentId }: TaxaTreemapProps) {
       });
     }
 
-    content = (
+    return (
       <ResponsiveContainer width="100%" height={320}>
         <Treemap
           data={treemapData}
@@ -118,87 +100,9 @@ function TaxaTreemap({ deploymentId }: TaxaTreemapProps) {
         </Treemap>
       </ResponsiveContainer>
     );
-  }
+  };
 
-  return (
-    <Card className="w-full self-start p-4 lg:w-1/2">
-      <CardHeader>
-        <CardTitle className="text-lg capitalize">Taxa treemap</CardTitle>
-        <CardDescription>Detection count by selected taxonomy level</CardDescription>
-      </CardHeader>
-      <Separator />
-      <CardContent>{content}</CardContent>
-    </Card>
-  );
-}
-
-function TaxaTreemapCell({
-  x = 0,
-  y = 0,
-  width = 0,
-  height = 0,
-  index = 0,
-  name,
-  count,
-  payload,
-}: {
-  x?: number;
-  y?: number;
-  width?: number;
-  height?: number;
-  index?: number;
-  name?: string;
-  count?: number;
-  payload?: TreemapItem;
-}) {
-  if (width <= 0 || height <= 0) {
-    return null;
-  }
-
-  const safeName =
-    typeof name === "string" ? name : typeof payload?.name === "string" ? payload.name : "";
-  const fill =
-    safeName === "Others" ? OTHER_TREEMAP_COLOR : TREEMAP_COLORS[index % TREEMAP_COLORS.length];
-  const detectionCount = count ?? payload?.count ?? 0;
-  const showName = width >= MIN_LABEL_WIDTH && height >= MIN_LABEL_HEIGHT;
-  const showCount = width >= MIN_COUNT_WIDTH && height >= MIN_COUNT_HEIGHT;
-  const fontSize = width >= 96 && height >= 52 ? PRIMARY_FONT_SIZE : SECONDARY_FONT_SIZE;
-  const label = showName ? truncateLabel(safeName, width, fontSize) : "";
-
-  return (
-    <g>
-      <rect
-        x={x}
-        y={y}
-        width={width}
-        height={height}
-        fill={fill}
-        fillOpacity={0.9}
-        stroke="rgba(255,255,255,0.08)"
-      />
-      {label && (
-        <text
-          x={x + CELL_PADDING_X}
-          y={y + 18}
-          fill="var(--color-background)"
-          fontSize={fontSize}
-          fontWeight={600}
-        >
-          {label}
-        </text>
-      )}
-      {showCount && (
-        <text
-          x={x + CELL_PADDING_X}
-          y={y + (label ? 34 : 20)}
-          fill="rgba(11,15,12,0.7)"
-          fontSize={10}
-        >
-          {detectionCount} detections
-        </text>
-      )}
-    </g>
-  );
+  return content();
 }
 
 function TaxaTreemapTooltip({
